+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+23/04/2022
+* added inputer: replace dashes with spaces
 
 14/07/2021
 * added inputer: REPLACE WORD WITH CAMELCASE TAG